Id tipovi refaktorisani u Long

+++ projekat/src/main/java/com/projekat/projekat_mrs_isa/repository/BoatRepository.java
@@ -3,5 +3,5 @@ package com.projekat.projekat_mrs_isa.repository;
 import com.projekat.projekat_mrs_isa.model.Boat;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoatRepository extends JpaRepository<Boat, Integer> {
+public interface BoatRepository extends JpaRepository<Boat, Long> {
 }
