+++ projekat/src/main/java/com/projekat/projekat_mrs_isa/repository/FishingClassRepository.java
@@ -3,5 +3,5 @@ package com.projekat.projekat_mrs_isa.repository;
 import com.projekat.projekat_mrs_isa.model.FishingClass;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FishingClassRepository extends JpaRepository<FishingClass, Integer> {
+public interface FishingClassRepository extends JpaRepository<FishingClass, Long> {
 }
