+++ projekat/src/main/java/com/projekat/projekat_mrs_isa/repository/FishingInstructorRepository.java
@@ -3,5 +3,5 @@ package com.projekat.projekat_mrs_isa.repository;
 import com.projekat.projekat_mrs_isa.model.FishingInstructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FishingInstructorRepository extends JpaRepository<FishingInstructor,Integer> {
+public interface FishingInstructorRepository extends JpaRepository<FishingInstructor,Long> {
 }
