+++ projekat/src/main/java/com/projekat/projekat_mrs_isa/repository/ReservationRepository.java
@@ -3,5 +3,5 @@ package com.projekat.projekat_mrs_isa.repository;
 import com.projekat.projekat_mrs_isa.model.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
 }
