+++ projekat/src/main/java/com/projekat/projekat_mrs_isa/repository/ShipOwnerRepository.java
@@ -4,5 +4,5 @@ package com.projekat.projekat_mrs_isa.repository;
 import com.projekat.projekat_mrs_isa.model.ShipOwner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShipOwnerRepository  extends JpaRepository<ShipOwner,Integer> {
+public interface ShipOwnerRepository  extends JpaRepository<ShipOwner,Long> {
 }
