+++ projekat/src/main/java/com/projekat/projekat_mrs_isa/repository/VacationHouseOwnerRepository.java
@@ -4,5 +4,5 @@ package com.projekat.projekat_mrs_isa.repository;
 import com.projekat.projekat_mrs_isa.model.VacationHouseOwner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VacationHouseOwnerRepository  extends JpaRepository<VacationHouseOwner,Integer> {
+public interface VacationHouseOwnerRepository  extends JpaRepository<VacationHouseOwner,Long> {
 }
