+++ projekat/src/main/java/com/projekat/projekat_mrs_isa/repository/VacationHouseRepository.java
@@ -3,5 +3,5 @@ package com.projekat.projekat_mrs_isa.repository;
 import com.projekat.projekat_mrs_isa.model.VacationHouse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VacationHouseRepository extends JpaRepository<VacationHouse, Integer> {
+public interface VacationHouseRepository extends JpaRepository<VacationHouse, Long> {
 }
